Rename files to use UpperCamelCaseCluster/Logic for ethernet-network-diagnostics-server
