+++ src/app/clusters/ethernet-network-diagnostics-server/EthernetDiagnosticsCluster.h
